This branch adds the persistent pt-to-pt test
